custom_resources/requirements.txt
boto3
python-dotenv
snowflake-snowpark-python
crhelper
Jinja2